+++ templates/home-manager/flake.nix
@@ -19,6 +19,7 @@
   outputs = { self, nixpkgs, home-manager, nix-on-droid }: {
 
     nixOnDroidConfigurations.default = nix-on-droid.lib.nixOnDroidConfiguration {
+      pkgs = import nixpkgs { system = "aarch64-linux"; };
       modules = [ ./nix-on-droid.nix ];
     };